cast to sizet

+++ src/outputs/DMXOutput.cpp
@@ -31,7 +31,7 @@ void DMXOutput::OutputFrame(uint8_t* data)
 {
     if (!Enabled || m_SerialPort == nullptr || m_SerialPort->isOpen()) return;
 
-    size_t chs = std::min(Channels, (size_t)(DMX_MAX_CHANNELS));
+    size_t chs = std::min((size_t)Channels, (size_t)(DMX_MAX_CHANNELS));
 
     if (memcmp(&_data[5], &data[StartChannel - 1], chs) == 0) {
         // nothing changed
