+++ src/outputs/GenericSerialOutput.cpp
@@ -40,7 +40,7 @@ void GenericSerialOutput::OutputFrame(uint8_t* data)
 {
     if (!Enabled || m_SerialPort == nullptr || m_SerialPort->isOpen()) return;
 
-    size_t chs = std::min(Channels, (size_t)(GENERICSERIAL_MAX_CHANNELS));
+    size_t chs = std::min((size_t)Channels, (size_t)(GENERICSERIAL_MAX_CHANNELS));
     if (memcmp(&_data[_prefix.size()], &data[StartChannel - 1], chs) != 0) {
         memcpy(&_data[_prefix.size()], &data[StartChannel - 1], chs);
     }
